+++ frontend/sdk/so_insights_client/api/analysis_runs/create_analysis_run.py
@@ -5,17 +5,16 @@ import httpx
 
 from ... import errors
 from ...client import AuthenticatedClient, Client
-from ...models.cluster import Cluster
-from ...models.cluster_feedback import ClusterFeedback
+from ...models.analysis_run import AnalysisRun
+from ...models.analysis_run_create import AnalysisRunCreate
 from ...models.http_validation_error import HTTPValidationError
 from ...types import UNSET, Response, Unset
 
 
 def _get_kwargs(
     workspace_id: str,
-    cluster_id: str,
     *,
-    body: ClusterFeedback,
+    body: AnalysisRunCreate,
     x_organization_id: Union[None, Unset, str] = UNSET,
 ) -> dict[str, Any]:
     headers: dict[str, Any] = {}
@@ -23,8 +22,8 @@ def _get_kwargs(
         headers["x-organization-id"] = x_organization_id
 
     _kwargs: dict[str, Any] = {
-        "method": "put",
-        "url": f"/workspaces/{workspace_id}/clustering/clusters/{cluster_id}/feedback",
+        "method": "post",
+        "url": f"/workspaces/{workspace_id}/analysis-runs/",
     }
 
     _body = body.to_dict()
@@ -38,9 +37,9 @@ def _get_kwargs(
 
 def _parse_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Union[Cluster, HTTPValidationError]]:
+) -> Optional[Union[AnalysisRun, HTTPValidationError]]:
     if response.status_code == 200:
-        response_200 = Cluster.from_dict(response.json())
+        response_200 = AnalysisRun.from_dict(response.json())
 
         return response_200
     if response.status_code == 422:
@@ -55,7 +54,7 @@ def _parse_response(
 
 def _build_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Union[Cluster, HTTPValidationError]]:
+) -> Response[Union[AnalysisRun, HTTPValidationError]]:
     return Response(
         status_code=HTTPStatus(response.status_code),
         content=response.content,
@@ -66,39 +65,30 @@ def _build_response(
 
 def sync_detailed(
     workspace_id: str,
-    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    body: ClusterFeedback,
+    body: AnalysisRunCreate,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[Cluster, HTTPValidationError]]:
-    """Set Cluster Feedback
+) -> Response[Union[AnalysisRun, HTTPValidationError]]:
+    """Create Analysis Run
 
-     Set or update feedback for a specific cluster
+     Create a new pending analysis run
 
     Args:
         workspace_id (str):
-        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
-        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
-            cluster.
-
-            This simple model allows users to indicate whether they find a particular
-            cluster of articles relevant or not. It's a way to incorporate human judgment
-            into the system's organization of content, which can help improve the
-            cluster evaluation process over time.
+        body (AnalysisRunCreate):
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[Cluster, HTTPValidationError]]
+        Response[Union[AnalysisRun, HTTPValidationError]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        cluster_id=cluster_id,
         body=body,
         x_organization_id=x_organization_id,
     )
@@ -112,39 +102,30 @@ def sync_detailed(
 
 def sync(
     workspace_id: str,
-    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    body: ClusterFeedback,
+    body: AnalysisRunCreate,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[Cluster, HTTPValidationError]]:
-    """Set Cluster Feedback
+) -> Optional[Union[AnalysisRun, HTTPValidationError]]:
+    """Create Analysis Run
 
-     Set or update feedback for a specific cluster
+     Create a new pending analysis run
 
     Args:
         workspace_id (str):
-        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
-        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
-            cluster.
-
-            This simple model allows users to indicate whether they find a particular
-            cluster of articles relevant or not. It's a way to incorporate human judgment
-            into the system's organization of content, which can help improve the
-            cluster evaluation process over time.
+        body (AnalysisRunCreate):
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[Cluster, HTTPValidationError]
+        Union[AnalysisRun, HTTPValidationError]
     """
 
     return sync_detailed(
         workspace_id=workspace_id,
-        cluster_id=cluster_id,
         client=client,
         body=body,
         x_organization_id=x_organization_id,
@@ -153,39 +134,30 @@ def sync(
 
 async def asyncio_detailed(
     workspace_id: str,
-    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    body: ClusterFeedback,
+    body: AnalysisRunCreate,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[Cluster, HTTPValidationError]]:
-    """Set Cluster Feedback
+) -> Response[Union[AnalysisRun, HTTPValidationError]]:
+    """Create Analysis Run
 
-     Set or update feedback for a specific cluster
+     Create a new pending analysis run
 
     Args:
         workspace_id (str):
-        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
-        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
-            cluster.
-
-            This simple model allows users to indicate whether they find a particular
-            cluster of articles relevant or not. It's a way to incorporate human judgment
-            into the system's organization of content, which can help improve the
-            cluster evaluation process over time.
+        body (AnalysisRunCreate):
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[Cluster, HTTPValidationError]]
+        Response[Union[AnalysisRun, HTTPValidationError]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        cluster_id=cluster_id,
         body=body,
         x_organization_id=x_organization_id,
     )
@@ -197,40 +169,31 @@ async def asyncio_detailed(
 
 async def asyncio(
     workspace_id: str,
-    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    body: ClusterFeedback,
+    body: AnalysisRunCreate,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[Cluster, HTTPValidationError]]:
-    """Set Cluster Feedback
+) -> Optional[Union[AnalysisRun, HTTPValidationError]]:
+    """Create Analysis Run
 
-     Set or update feedback for a specific cluster
+     Create a new pending analysis run
 
     Args:
         workspace_id (str):
-        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
-        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
-            cluster.
-
-            This simple model allows users to indicate whether they find a particular
-            cluster of articles relevant or not. It's a way to incorporate human judgment
-            into the system's organization of content, which can help improve the
-            cluster evaluation process over time.
+        body (AnalysisRunCreate):
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[Cluster, HTTPValidationError]
+        Union[AnalysisRun, HTTPValidationError]
     """
 
     return (
         await asyncio_detailed(
             workspace_id=workspace_id,
-            cluster_id=cluster_id,
             client=client,
             body=body,
             x_organization_id=x_organization_id,
