+++ frontend/sdk/so_insights_client/api/analysis_runs/delete_cluster_feedback.py
@@ -5,40 +5,24 @@ import httpx
 
 from ... import errors
 from ...client import AuthenticatedClient, Client
-from ...models.cluster_with_articles import ClusterWithArticles
+from ...models.cluster import Cluster
 from ...models.http_validation_error import HTTPValidationError
-from ...models.relevancy_filter import RelevancyFilter
 from ...types import UNSET, Response, Unset
 
 
 def _get_kwargs(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
-    relevancy_filter: Union[Unset, RelevancyFilter] = UNSET,
-    n_articles: Union[Unset, int] = 5,
     x_organization_id: Union[None, Unset, str] = UNSET,
 ) -> dict[str, Any]:
     headers: dict[str, Any] = {}
     if not isinstance(x_organization_id, Unset):
         headers["x-organization-id"] = x_organization_id
 
-    params: dict[str, Any] = {}
-
-    json_relevancy_filter: Union[Unset, str] = UNSET
-    if not isinstance(relevancy_filter, Unset):
-        json_relevancy_filter = relevancy_filter.value
-
-    params["relevancy_filter"] = json_relevancy_filter
-
-    params["n_articles"] = n_articles
-
-    params = {k: v for k, v in params.items() if v is not UNSET and v is not None}
-
     _kwargs: dict[str, Any] = {
-        "method": "get",
-        "url": f"/workspaces/{workspace_id}/clustering/sessions/{session_id}/clusters-with-articles",
-        "params": params,
+        "method": "delete",
+        "url": f"/workspaces/{workspace_id}/analysis-runs/clusters/{cluster_id}/feedback",
     }
 
     _kwargs["headers"] = headers
@@ -47,14 +31,9 @@ def _get_kwargs(
 
 def _parse_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Union[HTTPValidationError, list["ClusterWithArticles"]]]:
+) -> Optional[Union[Cluster, HTTPValidationError]]:
     if response.status_code == 200:
-        response_200 = []
-        _response_200 = response.json()
-        for response_200_item_data in _response_200:
-            response_200_item = ClusterWithArticles.from_dict(response_200_item_data)
-
-            response_200.append(response_200_item)
+        response_200 = Cluster.from_dict(response.json())
 
         return response_200
     if response.status_code == 422:
@@ -69,7 +48,7 @@ def _parse_response(
 
 def _build_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Union[HTTPValidationError, list["ClusterWithArticles"]]]:
+) -> Response[Union[Cluster, HTTPValidationError]]:
     return Response(
         status_code=HTTPStatus(response.status_code),
         content=response.content,
@@ -80,22 +59,18 @@ def _build_response(
 
 def sync_detailed(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevancy_filter: Union[Unset, RelevancyFilter] = UNSET,
-    n_articles: Union[Unset, int] = 5,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[HTTPValidationError, list["ClusterWithArticles"]]]:
-    """List Clusters With Articles
+) -> Response[Union[Cluster, HTTPValidationError]]:
+    """Delete Cluster Feedback
 
-     List all clusters with their top N articles for a specific clustering session
+     Delete feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevancy_filter (Union[Unset, RelevancyFilter]):
-        n_articles (Union[Unset, int]):  Default: 5.
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -103,14 +78,12 @@ def sync_detailed(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[HTTPValidationError, list['ClusterWithArticles']]]
+        Response[Union[Cluster, HTTPValidationError]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        session_id=session_id,
-        relevancy_filter=relevancy_filter,
-        n_articles=n_articles,
+        cluster_id=cluster_id,
         x_organization_id=x_organization_id,
     )
 
@@ -123,22 +96,18 @@ def sync_detailed(
 
 def sync(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevancy_filter: Union[Unset, RelevancyFilter] = UNSET,
-    n_articles: Union[Unset, int] = 5,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[HTTPValidationError, list["ClusterWithArticles"]]]:
-    """List Clusters With Articles
+) -> Optional[Union[Cluster, HTTPValidationError]]:
+    """Delete Cluster Feedback
 
-     List all clusters with their top N articles for a specific clustering session
+     Delete feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevancy_filter (Union[Unset, RelevancyFilter]):
-        n_articles (Union[Unset, int]):  Default: 5.
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -146,37 +115,31 @@ def sync(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[HTTPValidationError, list['ClusterWithArticles']]
+        Union[Cluster, HTTPValidationError]
     """
 
     return sync_detailed(
         workspace_id=workspace_id,
-        session_id=session_id,
+        cluster_id=cluster_id,
         client=client,
-        relevancy_filter=relevancy_filter,
-        n_articles=n_articles,
         x_organization_id=x_organization_id,
     ).parsed
 
 
 async def asyncio_detailed(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevancy_filter: Union[Unset, RelevancyFilter] = UNSET,
-    n_articles: Union[Unset, int] = 5,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[HTTPValidationError, list["ClusterWithArticles"]]]:
-    """List Clusters With Articles
+) -> Response[Union[Cluster, HTTPValidationError]]:
+    """Delete Cluster Feedback
 
-     List all clusters with their top N articles for a specific clustering session
+     Delete feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevancy_filter (Union[Unset, RelevancyFilter]):
-        n_articles (Union[Unset, int]):  Default: 5.
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -184,14 +147,12 @@ async def asyncio_detailed(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[HTTPValidationError, list['ClusterWithArticles']]]
+        Response[Union[Cluster, HTTPValidationError]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        session_id=session_id,
-        relevancy_filter=relevancy_filter,
-        n_articles=n_articles,
+        cluster_id=cluster_id,
         x_organization_id=x_organization_id,
     )
 
@@ -202,22 +163,18 @@ async def asyncio_detailed(
 
 async def asyncio(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevancy_filter: Union[Unset, RelevancyFilter] = UNSET,
-    n_articles: Union[Unset, int] = 5,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[HTTPValidationError, list["ClusterWithArticles"]]]:
-    """List Clusters With Articles
+) -> Optional[Union[Cluster, HTTPValidationError]]:
+    """Delete Cluster Feedback
 
-     List all clusters with their top N articles for a specific clustering session
+     Delete feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevancy_filter (Union[Unset, RelevancyFilter]):
-        n_articles (Union[Unset, int]):  Default: 5.
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -225,16 +182,14 @@ async def asyncio(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[HTTPValidationError, list['ClusterWithArticles']]
+        Union[Cluster, HTTPValidationError]
     """
 
     return (
         await asyncio_detailed(
             workspace_id=workspace_id,
-            session_id=session_id,
+            cluster_id=cluster_id,
             client=client,
-            relevancy_filter=relevancy_filter,
-            n_articles=n_articles,
             x_organization_id=x_organization_id,
         )
     ).parsed
