+++ frontend/sdk/so_insights_client/api/analysis_runs/list_clusters_for_clustering_run.py
@@ -5,24 +5,46 @@ import httpx
 
 from ... import errors
 from ...client import AuthenticatedClient, Client
-from ...models.clustering_session import ClusteringSession
+from ...models.cluster import Cluster
 from ...models.http_validation_error import HTTPValidationError
+from ...models.list_clusters_for_clustering_run_relevance_levels_type_0_item import (
+    ListClustersForClusteringRunRelevanceLevelsType0Item,
+)
 from ...types import UNSET, Response, Unset
 
 
 def _get_kwargs(
     workspace_id: str,
-    session_id: str,
+    analysis_run_id: str,
     *,
+    relevance_levels: Union[None, Unset, list[ListClustersForClusteringRunRelevanceLevelsType0Item]] = UNSET,
     x_organization_id: Union[None, Unset, str] = UNSET,
 ) -> dict[str, Any]:
     headers: dict[str, Any] = {}
     if not isinstance(x_organization_id, Unset):
         headers["x-organization-id"] = x_organization_id
 
+    params: dict[str, Any] = {}
+
+    json_relevance_levels: Union[None, Unset, list[str]]
+    if isinstance(relevance_levels, Unset):
+        json_relevance_levels = UNSET
+    elif isinstance(relevance_levels, list):
+        json_relevance_levels = []
+        for relevance_levels_type_0_item_data in relevance_levels:
+            relevance_levels_type_0_item = relevance_levels_type_0_item_data.value
+            json_relevance_levels.append(relevance_levels_type_0_item)
+
+    else:
+        json_relevance_levels = relevance_levels
+    params["relevance_levels"] = json_relevance_levels
+
+    params = {k: v for k, v in params.items() if v is not UNSET and v is not None}
+
     _kwargs: dict[str, Any] = {
         "method": "get",
-        "url": f"/workspaces/{workspace_id}/clustering/sessions/{session_id}",
+        "url": f"/workspaces/{workspace_id}/analysis-runs/{analysis_run_id}/clusters",
+        "params": params,
     }
 
     _kwargs["headers"] = headers
@@ -31,9 +53,14 @@ def _get_kwargs(
 
 def _parse_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Union[ClusteringSession, HTTPValidationError]]:
+) -> Optional[Union[HTTPValidationError, list["Cluster"]]]:
     if response.status_code == 200:
-        response_200 = ClusteringSession.from_dict(response.json())
+        response_200 = []
+        _response_200 = response.json()
+        for response_200_item_data in _response_200:
+            response_200_item = Cluster.from_dict(response_200_item_data)
+
+            response_200.append(response_200_item)
 
         return response_200
     if response.status_code == 422:
@@ -48,7 +75,7 @@ def _parse_response(
 
 def _build_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Union[ClusteringSession, HTTPValidationError]]:
+) -> Response[Union[HTTPValidationError, list["Cluster"]]]:
     return Response(
         status_code=HTTPStatus(response.status_code),
         content=response.content,
@@ -59,18 +86,21 @@ def _build_response(
 
 def sync_detailed(
     workspace_id: str,
-    session_id: str,
+    analysis_run_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
+    relevance_levels: Union[None, Unset, list[ListClustersForClusteringRunRelevanceLevelsType0Item]] = UNSET,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[ClusteringSession, HTTPValidationError]]:
-    """Get Clustering Session
+) -> Response[Union[HTTPValidationError, list["Cluster"]]]:
+    """List Clusters
 
-     Get a specific clustering session
+     List all clusters for a specific analysis run
 
     Args:
         workspace_id (str):
-        session_id (str):
+        analysis_run_id (str):
+        relevance_levels (Union[None, Unset,
+            list[ListClustersForClusteringRunRelevanceLevelsType0Item]]):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -78,12 +108,13 @@ def sync_detailed(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[ClusteringSession, HTTPValidationError]]
+        Response[Union[HTTPValidationError, list['Cluster']]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        session_id=session_id,
+        analysis_run_id=analysis_run_id,
+        relevance_levels=relevance_levels,
         x_organization_id=x_organization_id,
     )
 
@@ -96,18 +127,21 @@ def sync_detailed(
 
 def sync(
     workspace_id: str,
-    session_id: str,
+    analysis_run_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
+    relevance_levels: Union[None, Unset, list[ListClustersForClusteringRunRelevanceLevelsType0Item]] = UNSET,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[ClusteringSession, HTTPValidationError]]:
-    """Get Clustering Session
+) -> Optional[Union[HTTPValidationError, list["Cluster"]]]:
+    """List Clusters
 
-     Get a specific clustering session
+     List all clusters for a specific analysis run
 
     Args:
         workspace_id (str):
-        session_id (str):
+        analysis_run_id (str):
+        relevance_levels (Union[None, Unset,
+            list[ListClustersForClusteringRunRelevanceLevelsType0Item]]):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -115,31 +149,35 @@ def sync(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[ClusteringSession, HTTPValidationError]
+        Union[HTTPValidationError, list['Cluster']]
     """
 
     return sync_detailed(
         workspace_id=workspace_id,
-        session_id=session_id,
+        analysis_run_id=analysis_run_id,
         client=client,
+        relevance_levels=relevance_levels,
         x_organization_id=x_organization_id,
     ).parsed
 
 
 async def asyncio_detailed(
     workspace_id: str,
-    session_id: str,
+    analysis_run_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
+    relevance_levels: Union[None, Unset, list[ListClustersForClusteringRunRelevanceLevelsType0Item]] = UNSET,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[ClusteringSession, HTTPValidationError]]:
-    """Get Clustering Session
+) -> Response[Union[HTTPValidationError, list["Cluster"]]]:
+    """List Clusters
 
-     Get a specific clustering session
+     List all clusters for a specific analysis run
 
     Args:
         workspace_id (str):
-        session_id (str):
+        analysis_run_id (str):
+        relevance_levels (Union[None, Unset,
+            list[ListClustersForClusteringRunRelevanceLevelsType0Item]]):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -147,12 +185,13 @@ async def asyncio_detailed(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[ClusteringSession, HTTPValidationError]]
+        Response[Union[HTTPValidationError, list['Cluster']]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        session_id=session_id,
+        analysis_run_id=analysis_run_id,
+        relevance_levels=relevance_levels,
         x_organization_id=x_organization_id,
     )
 
@@ -163,18 +202,21 @@ async def asyncio_detailed(
 
 async def asyncio(
     workspace_id: str,
-    session_id: str,
+    analysis_run_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
+    relevance_levels: Union[None, Unset, list[ListClustersForClusteringRunRelevanceLevelsType0Item]] = UNSET,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[ClusteringSession, HTTPValidationError]]:
-    """Get Clustering Session
+) -> Optional[Union[HTTPValidationError, list["Cluster"]]]:
+    """List Clusters
 
-     Get a specific clustering session
+     List all clusters for a specific analysis run
 
     Args:
         workspace_id (str):
-        session_id (str):
+        analysis_run_id (str):
+        relevance_levels (Union[None, Unset,
+            list[ListClustersForClusteringRunRelevanceLevelsType0Item]]):
         x_organization_id (Union[None, Unset, str]):
 
     Raises:
@@ -182,14 +224,15 @@ async def asyncio(
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[ClusteringSession, HTTPValidationError]
+        Union[HTTPValidationError, list['Cluster']]
     """
 
     return (
         await asyncio_detailed(
             workspace_id=workspace_id,
-            session_id=session_id,
+            analysis_run_id=analysis_run_id,
             client=client,
+            relevance_levels=relevance_levels,
             x_organization_id=x_organization_id,
         )
     ).parsed
