+++ frontend/sdk/so_insights_client/api/analysis_runs/set_cluster_feedback.py
@@ -6,46 +6,31 @@ import httpx
 from ... import errors
 from ...client import AuthenticatedClient, Client
 from ...models.cluster import Cluster
+from ...models.cluster_feedback import ClusterFeedback
 from ...models.http_validation_error import HTTPValidationError
-from ...models.list_clusters_for_session_relevance_levels_type_0_item import (
-    ListClustersForSessionRelevanceLevelsType0Item,
-)
 from ...types import UNSET, Response, Unset
 
 
 def _get_kwargs(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
-    relevance_levels: Union[None, Unset, list[ListClustersForSessionRelevanceLevelsType0Item]] = UNSET,
+    body: ClusterFeedback,
     x_organization_id: Union[None, Unset, str] = UNSET,
 ) -> dict[str, Any]:
     headers: dict[str, Any] = {}
     if not isinstance(x_organization_id, Unset):
         headers["x-organization-id"] = x_organization_id
 
-    params: dict[str, Any] = {}
-
-    json_relevance_levels: Union[None, Unset, list[str]]
-    if isinstance(relevance_levels, Unset):
-        json_relevance_levels = UNSET
-    elif isinstance(relevance_levels, list):
-        json_relevance_levels = []
-        for relevance_levels_type_0_item_data in relevance_levels:
-            relevance_levels_type_0_item = relevance_levels_type_0_item_data.value
-            json_relevance_levels.append(relevance_levels_type_0_item)
-
-    else:
-        json_relevance_levels = relevance_levels
-    params["relevance_levels"] = json_relevance_levels
-
-    params = {k: v for k, v in params.items() if v is not UNSET and v is not None}
-
     _kwargs: dict[str, Any] = {
-        "method": "get",
-        "url": f"/workspaces/{workspace_id}/clustering/sessions/{session_id}/clusters",
-        "params": params,
+        "method": "put",
+        "url": f"/workspaces/{workspace_id}/analysis-runs/clusters/{cluster_id}/feedback",
     }
+
+    _body = body.to_dict()
+
+    _kwargs["json"] = _body
+    headers["Content-Type"] = "application/json"
 
     _kwargs["headers"] = headers
     return _kwargs
@@ -53,14 +38,9 @@ def _get_kwargs(
 
 def _parse_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Optional[Union[HTTPValidationError, list["Cluster"]]]:
+) -> Optional[Union[Cluster, HTTPValidationError]]:
     if response.status_code == 200:
-        response_200 = []
-        _response_200 = response.json()
-        for response_200_item_data in _response_200:
-            response_200_item = Cluster.from_dict(response_200_item_data)
-
-            response_200.append(response_200_item)
+        response_200 = Cluster.from_dict(response.json())
 
         return response_200
     if response.status_code == 422:
@@ -75,7 +55,7 @@ def _parse_response(
 
 def _build_response(
     *, client: Union[AuthenticatedClient, Client], response: httpx.Response
-) -> Response[Union[HTTPValidationError, list["Cluster"]]]:
+) -> Response[Union[Cluster, HTTPValidationError]]:
     return Response(
         status_code=HTTPStatus(response.status_code),
         content=response.content,
@@ -86,35 +66,40 @@ def _build_response(
 
 def sync_detailed(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevance_levels: Union[None, Unset, list[ListClustersForSessionRelevanceLevelsType0Item]] = UNSET,
+    body: ClusterFeedback,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[HTTPValidationError, list["Cluster"]]]:
-    """List Clusters
+) -> Response[Union[Cluster, HTTPValidationError]]:
+    """Set Cluster Feedback
 
-     List all clusters for a specific clustering session
+     Set or update feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevance_levels (Union[None, Unset,
-            list[ListClustersForSessionRelevanceLevelsType0Item]]):
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
+        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
+            cluster.
+
+            This simple model allows users to indicate whether they find a particular
+            cluster of articles relevant or not. It's a way to incorporate human judgment
+            into the system's organization of content, which can help improve the
+            cluster evaluation process over time.
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[HTTPValidationError, list['Cluster']]]
+        Response[Union[Cluster, HTTPValidationError]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        session_id=session_id,
-        relevance_levels=relevance_levels,
+        cluster_id=cluster_id,
+        body=body,
         x_organization_id=x_organization_id,
     )
 
@@ -127,71 +112,81 @@ def sync_detailed(
 
 def sync(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevance_levels: Union[None, Unset, list[ListClustersForSessionRelevanceLevelsType0Item]] = UNSET,
+    body: ClusterFeedback,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[HTTPValidationError, list["Cluster"]]]:
-    """List Clusters
+) -> Optional[Union[Cluster, HTTPValidationError]]:
+    """Set Cluster Feedback
 
-     List all clusters for a specific clustering session
+     Set or update feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevance_levels (Union[None, Unset,
-            list[ListClustersForSessionRelevanceLevelsType0Item]]):
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
+        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
+            cluster.
+
+            This simple model allows users to indicate whether they find a particular
+            cluster of articles relevant or not. It's a way to incorporate human judgment
+            into the system's organization of content, which can help improve the
+            cluster evaluation process over time.
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[HTTPValidationError, list['Cluster']]
+        Union[Cluster, HTTPValidationError]
     """
 
     return sync_detailed(
         workspace_id=workspace_id,
-        session_id=session_id,
+        cluster_id=cluster_id,
         client=client,
-        relevance_levels=relevance_levels,
+        body=body,
         x_organization_id=x_organization_id,
     ).parsed
 
 
 async def asyncio_detailed(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevance_levels: Union[None, Unset, list[ListClustersForSessionRelevanceLevelsType0Item]] = UNSET,
+    body: ClusterFeedback,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Response[Union[HTTPValidationError, list["Cluster"]]]:
-    """List Clusters
+) -> Response[Union[Cluster, HTTPValidationError]]:
+    """Set Cluster Feedback
 
-     List all clusters for a specific clustering session
+     Set or update feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevance_levels (Union[None, Unset,
-            list[ListClustersForSessionRelevanceLevelsType0Item]]):
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
+        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
+            cluster.
+
+            This simple model allows users to indicate whether they find a particular
+            cluster of articles relevant or not. It's a way to incorporate human judgment
+            into the system's organization of content, which can help improve the
+            cluster evaluation process over time.
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Response[Union[HTTPValidationError, list['Cluster']]]
+        Response[Union[Cluster, HTTPValidationError]]
     """
 
     kwargs = _get_kwargs(
         workspace_id=workspace_id,
-        session_id=session_id,
-        relevance_levels=relevance_levels,
+        cluster_id=cluster_id,
+        body=body,
         x_organization_id=x_organization_id,
     )
 
@@ -202,37 +197,42 @@ async def asyncio_detailed(
 
 async def asyncio(
     workspace_id: str,
-    session_id: str,
+    cluster_id: str,
     *,
     client: Union[AuthenticatedClient, Client],
-    relevance_levels: Union[None, Unset, list[ListClustersForSessionRelevanceLevelsType0Item]] = UNSET,
+    body: ClusterFeedback,
     x_organization_id: Union[None, Unset, str] = UNSET,
-) -> Optional[Union[HTTPValidationError, list["Cluster"]]]:
-    """List Clusters
+) -> Optional[Union[Cluster, HTTPValidationError]]:
+    """Set Cluster Feedback
 
-     List all clusters for a specific clustering session
+     Set or update feedback for a specific cluster
 
     Args:
         workspace_id (str):
-        session_id (str):
-        relevance_levels (Union[None, Unset,
-            list[ListClustersForSessionRelevanceLevelsType0Item]]):
+        cluster_id (str):
         x_organization_id (Union[None, Unset, str]):
+        body (ClusterFeedback): Captures user feedback on the relevance or usefulness of a
+            cluster.
+
+            This simple model allows users to indicate whether they find a particular
+            cluster of articles relevant or not. It's a way to incorporate human judgment
+            into the system's organization of content, which can help improve the
+            cluster evaluation process over time.
 
     Raises:
         errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
         httpx.TimeoutException: If the request takes longer than Client.timeout.
 
     Returns:
-        Union[HTTPValidationError, list['Cluster']]
+        Union[Cluster, HTTPValidationError]
     """
 
     return (
         await asyncio_detailed(
             workspace_id=workspace_id,
-            session_id=session_id,
+            cluster_id=cluster_id,
             client=client,
-            relevance_levels=relevance_levels,
+            body=body,
             x_organization_id=x_organization_id,
         )
     ).parsed
