+++ frontend/sdk/so_insights_client/models/__init__.py
@@ -1,7 +1,15 @@
 """Contains all the data models used in inputs/outputs"""
 
+from .analysis_result import AnalysisResult
+from .analysis_result_analysis_type import AnalysisResultAnalysisType
+from .analysis_run import AnalysisRun
+from .analysis_run_analysis_type import AnalysisRunAnalysisType
+from .analysis_run_create import AnalysisRunCreate
+from .analysis_run_create_analysis_type import AnalysisRunCreateAnalysisType
 from .article import Article
 from .article_content_cleaner_output import ArticleContentCleanerOutput
+from .article_evaluation import ArticleEvaluation
+from .article_evaluation_relevance_level import ArticleEvaluationRelevanceLevel
 from .article_preview import ArticlePreview
 from .cluster import Cluster
 from .cluster_evaluation import ClusterEvaluation
@@ -9,9 +17,7 @@ from .cluster_evaluation_relevance_level import ClusterEvaluationRelevanceLevel
 from .cluster_feedback import ClusterFeedback
 from .cluster_overview import ClusterOverview
 from .cluster_with_articles import ClusterWithArticles
-from .clustering_session import ClusteringSession
-from .clustering_session_create import ClusteringSessionCreate
-from .clustering_session_metadata import ClusteringSessionMetadata
+from .clustering_analysis_params import ClusteringAnalysisParams
 from .content_fetching_result import ContentFetchingResult
 from .hdbscan_settings import HdbscanSettings
 from .http_validation_error import HTTPValidationError
@@ -20,11 +26,14 @@ from .ingestion_run import IngestionRun
 from .language import Language
 from .list_articles_sort_by import ListArticlesSortBy
 from .list_articles_sort_order import ListArticlesSortOrder
-from .list_clusters_for_session_relevance_levels_type_0_item import ListClustersForSessionRelevanceLevelsType0Item
+from .list_clusters_for_clustering_run_relevance_levels_type_0_item import (
+    ListClustersForClusteringRunRelevanceLevelsType0Item,
+)
 from .organization import Organization
 from .paginated_response_article import PaginatedResponseArticle
 from .region import Region
 from .relevancy_filter import RelevancyFilter
+from .report_analysis_params import ReportAnalysisParams
 from .rss_ingestion_config import RssIngestionConfig
 from .rss_ingestion_config_create import RssIngestionConfigCreate
 from .rss_ingestion_config_update import RssIngestionConfigUpdate
@@ -42,16 +51,22 @@ from .workspace_create import WorkspaceCreate
 from .workspace_update import WorkspaceUpdate
 
 __all__ = (
+    "AnalysisResult",
+    "AnalysisResultAnalysisType",
+    "AnalysisRun",
+    "AnalysisRunAnalysisType",
+    "AnalysisRunCreate",
+    "AnalysisRunCreateAnalysisType",
     "Article",
     "ArticleContentCleanerOutput",
+    "ArticleEvaluation",
+    "ArticleEvaluationRelevanceLevel",
     "ArticlePreview",
     "Cluster",
     "ClusterEvaluation",
     "ClusterEvaluationRelevanceLevel",
     "ClusterFeedback",
-    "ClusteringSession",
-    "ClusteringSessionCreate",
-    "ClusteringSessionMetadata",
+    "ClusteringAnalysisParams",
     "ClusterOverview",
     "ClusterWithArticles",
     "ContentFetchingResult",
@@ -62,11 +77,12 @@ __all__ = (
     "Language",
     "ListArticlesSortBy",
     "ListArticlesSortOrder",
-    "ListClustersForSessionRelevanceLevelsType0Item",
+    "ListClustersForClusteringRunRelevanceLevelsType0Item",
     "Organization",
     "PaginatedResponseArticle",
     "Region",
     "RelevancyFilter",
+    "ReportAnalysisParams",
     "RssIngestionConfig",
     "RssIngestionConfigCreate",
     "RssIngestionConfigUpdate",
