+++ frontend/sdk/so_insights_client/models/analysis_result_analysis_type.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class AnalysisResultAnalysisType(str, Enum):
+    CLUSTERING = "clustering"
+    REPORT = "report"
+
+    def __str__(self) -> str:
+        return str(self.value)
