+++ frontend/sdk/so_insights_client/models/analysis_run_create_analysis_type.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class AnalysisRunCreateAnalysisType(str, Enum):
+    CLUSTERING = "clustering"
+    REPORT = "report"
+
+    def __str__(self) -> str:
+        return str(self.value)
