+++ frontend/sdk/so_insights_client/models/article_evaluation_relevance_level.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class ArticleEvaluationRelevanceLevel(str, Enum):
+    NOT_RELEVANT = "not_relevant"
+    RELEVANT = "relevant"
+    SOMEWHAT_RELEVANT = "somewhat_relevant"
+
+    def __str__(self) -> str:
+        return str(self.value)
