+++ frontend/sdk/so_insights_client/models/list_clusters_for_clustering_run_relevance_levels_type_0_item.py
@@ -1,7 +1,7 @@
 from enum import Enum
 
 
-class ListClustersForSessionRelevanceLevelsType0Item(str, Enum):
+class ListClustersForClusteringRunRelevanceLevelsType0Item(str, Enum):
     HIGHLY_RELEVANT = "highly_relevant"
     NOT_RELEVANT = "not_relevant"
     SOMEWHAT_RELEVANT = "somewhat_relevant"
